refactor: [피드백 반영] exception 패키지명 오타 수정
