+++ src/main/java/racewinner/exception/InvalidDelimiterException.java
@@ -1,4 +1,4 @@
-package racewinner.expteion;
+package racewinner.exception;
 
 public class InvalidDelimiterException extends RuntimeException {
     private final static String INVALID_DELIMITER_MSG = "','구분자 외의 다른 구분자를 사용 할 수 없습니다.";
